backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis[asyncio]==5.0.4
httpx==0.27.0
twilio==9.0.4
openai==1.30.1
razorpay==1.4.1
pillow==10.3.0
python-magic==0.4.27
aiofiles==23.2.1
pandas==2.2.2
python-csv==0.0.13
slowapi==0.1.9
loguru==0.7.2
apscheduler==3.10.4
